[Switch] Fix API URI
Hello,

I just install the script and notice the switch command gives me an error. Debugging it with the `-d` option, I got a 404 on the `switch/status` URI. 

According to the [doc](https://dev.freebox.fr/sdk/os/switch/), it seems the trailing slash of the URI makes the difference.